moving qty code and plots to new file
